# storage/api-docs/api-docs-v1.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lang Cards API v1",
        "description": "Lang Cards API ",
        "contact": {
            "email": "vlad2000100600@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/registration": {
            "post": {
                "tags": [
                    "Регистрация"
                ],
                "summary": "Регистрация нового пользователя",
                "operationId": "registerUser",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/AcceptLanguageHeader"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegistrationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно зарегистрирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Пользователь успешно зарегистрирован"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации входных данных",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Поле email уже используется."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка при регистрации пользователя",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "enum": [
                                                "success",
                                                "error"
                                            ],
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Не удалось зарегистрировать пользователя."
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null,
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RegistrationRequest": {
                "title": "Registration Request (Данные для регистрации)",
                "description": "Данные, необходимые для регистрации нового пользователя",
                "required": [
                    "name",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 255,
                        "example": "ivan@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "minLength": 8,
                        "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]+$",
                        "example": "Str0ngP@ssword!"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "Str0ngP@ssword!"
                    }
                },
                "type": "object"
            },
...
        },
...
        "parameters": {
            "AcceptLanguageHeader": {
                "name": "Accept-Language",
                "in": "header",
                "description": "Язык локали",
                "required": true,
                "schema": {
                    "type": "string",
                    "default": "ru",
                    "enum": [
                        "ru",
                        "en"
                    ]
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}